requirements.txt
azure-ai-inference==1.0.0b9
streamlit==1.43.2